Worker.compute supports receiving a task
